backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "JupyterLab environment with DDSP and other dependencies"
authors = [
  { name = "takemotoooo-eeic", email = "takemoto-kengo2262@g.ecc.u-tokyo.ac.jp" }
]
readme = "README.md"
requires-python = ">=3.9.7,<3.13"
dependencies = [
  "librosa>=0.9.2",
  "tqdm>=4.67.1",
  "pydantic>=2.11.3",
  "effortless-config>=0.7.0",
  "torch>=2.6.0",
  "pyyaml==5.3.1",
  "flask==1.1.2",
  "numpy>=1.23.5,<2.0.0",
  "einops==0.3.0",
  "soundfile>=0.12.1",
  "crepe==0.0.16",
  "tensorflow==2.15.0",
  "ruff>=0.11.5",
  "python-dotenv>=1.1.0",
  "wandb>=0.19.9",
  "torchaudio>=2.6.0",
  "synctoolbox>=1.4.0",
  "lxml>=5.4.0",
  "transformers>=4.52.1",
  "imblearn>=0.0",
  "uvicorn>=0.34.3",
  "fastapi>=0.115.12",
  "pydantic-settings>=2.9.1",
  "python-multipart>=0.0.20",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff.lint.extend-per-file-ignores]
"api/app.py" = ["E402"]

[dependency-groups]
dev = [
    "ruff>=0.11.12",
]
